add buildAddressForLegalRepIccLetter
